# Use a stronger hash collision function