--- a/promql/engine.go
+++ b/promql/engine.go
@@ -1618,10 +1618,10 @@ func resultMetric(lhs, rhs labels.Labels, op ItemType, matching *VectorMatching,
 	}
 	// op and matching are always the same for a given node, so
 	// there's no need to include them in the hash key.
-	// If the lhs and rhs are the same then the xor would be 0,
-	// so add in one side to protect against that.
+	// The hash combine function is coming from the OpenGL Mathematics
+	// project and is licensed under MIT license.
 	lh := lhs.Hash()
-	h := (lh ^ rhs.Hash()) + lh
+	h := rhs.Hash() + 0x9e3779b9 + (lh << 6) + (lh >> 2)
 	if ret, ok := enh.resultMetric[h]; ok {
 		return ret
 	}