@roidelapluie 2020-01-04
Benchmark

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSignature-8          584           587           +0.51%
BenchmarkSignatureReuse-8     128           132           +3.12%

benchmark                     old allocs     new allocs     delta
BenchmarkSignature-8          7              7              +0.00%
BenchmarkSignatureReuse-8     1              1              +0.00%

benchmark                     old bytes     new bytes     delta
BenchmarkSignature-8          672           672           +0.00%
BenchmarkSignatureReuse-8     32            32            +0.00%

```